pyproject.toml
[tool.poetry]
name = "card_draw"
version = "0.1.0"
description = ""
authors = ["Axel Widenfelt <axel.widenfelt@esa.int>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"

[tool.poetry.scripts]
play = "card_draw.card_draw:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
